Refactor Keyed Ciphers to use common base class
