+++ package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "ajv": "^6.4.0",
     "bootstrap": "^3.3.7",
-    "puzzle-lib": "^0.1.2",
+    "puzzle-lib": "^0.1.4",
     "react": "^16.3.1",
     "react-bootstrap": "^0.32.1",
     "react-dom": "^16.3.1",
